+++ budget-alerts/package.json
@@ -15,6 +15,7 @@
         "notifications-node-client": "8.0.0"
     },
     "devDependencies": {
+        "axios-mock-adapter": "^1.22.0",
         "jest": "^29.7.0",
         "msw": "^2.3.0"
     }
