+++ taskfiles/budget-alerts.yaml
@@ -5,6 +5,11 @@ tasks:
     dir: budget-alerts
     cmds:
       - npx @google-cloud/functions-framework --target=manageBudgetAlerts
+  tests:
+    desc: Start budget-alerts cloud function tests
+    dir: budget-alerts
+    cmds:
+      - npm run test
   
   emulator:
     desc: Start local pubsub emulator
